+++ packages/backend/src/database/seeds/development/01_initial_user.ts
@@ -185,7 +185,7 @@ export async function seed(knex: Knex): Promise<void> {
             explores,
         );
     } catch (e) {
-        console.log(e);
+        console.error(e);
         throw e;
     }
 }
