+++ packages/common/src/types/api/share.ts
@@ -1,0 +1,6 @@
+import { ShareUrl } from '../share';
+
+export type ApiShareResponse = {
+    status: 'ok';
+    results: ShareUrl;
+};
