+++ packages/common/src/types/errors.ts
@@ -1,33 +1,6 @@
 // eslint-disable-next-line max-classes-per-file
 import { DbtLog } from './job';
 
-/**
- * The Error object is returned from the api any time there is an error.
- * The message contains
- */
-export type ApiErrorPayload = {
-    status: 'error';
-    error: {
-        /**
-         * HTTP status code
-         * @format integer
-         */
-        statusCode: number;
-        /**
-         * Unique name for the type of error
-         */
-        name: string;
-        /**
-         * A friendly message summarising the error
-         */
-        message?: string;
-        /**
-         * Optional data containing details of the error
-         */
-        data?: any;
-    };
-};
-
 type LightdashErrorParams = {
     message: string;
     name: string;
