+++ packages/common/src/types/share.ts
@@ -30,8 +30,3 @@ export type ShareUrl = {
  * Contains the detail of a full URL to generate a short URL id
  */
 export type CreateShareUrl = Pick<ShareUrl, 'path' | 'params'>;
-
-export type ApiShareResponse = {
-    status: 'ok';
-    results: ShareUrl;
-};
